requirements.txt
fastapi
uvicorn
pillow
numpy
onnxruntime
torch